# specs/001-furniture-quotation-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 家具報價單系統 API
  description: |
    家具報價單系統後端 API，提供 PDF 上傳、BOQ 解析、Excel 報價單產出功能。

    ## 功能概述
    - PDF 檔案上傳與管理
    - BOQ 表格自動解析（使用 Gemini AI）
    - 平面圖數量核對
    - 惠而蒙格式 Excel 報價單產出

    ## 錯誤處理
    所有錯誤回應皆使用繁體中文訊息。
  version: 1.0.0
  contact:
    name: API 支援
servers:
  - url: http://localhost:8000
    description: 開發環境
  - url: http://api.example.com
    description: 生產環境
...
paths:
...
  /api/documents/{document_id}/parsing:
    post:
      tags:
        - Parse
      summary: 啟動 PDF 解析
      description: |
        對已上傳的 PDF 文件啟動 BOQ 解析任務。
        解析使用 Gemini AI 進行，為非同步操作。
      operationId: startParsing
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                extract_images:
                  type: boolean
                  default: true
                  description: 是否提取圖片
                target_categories:
                  type: array
                  items:
                    type: string
                  description: 目標類別篩選（如：活動家具）
      responses:
        '202':
          description: 解析任務已建立
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: 請求無效
        '404':
          description: 找不到文件
  /api/documents/{document_id}/parse-result:
    get:
      tags:
        - Parse
      summary: 取得解析結果
      description: 取得 PDF 解析完成後的 BOQ 項目列表
      operationId: getParseResult
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  document_id:
                    type: string
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BOQItem'
                  images:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExtractedImage'
                  statistics:
                    $ref: '#/components/schemas/ParseStatistics'
        '404':
          description: 找不到文件或解析尚未完成
  /api/floor-plan-analyses:
    post:
      tags:
        - Parse
      summary: 平面圖數量核對
      description: |
        分析平面圖 PDF，核對並補充 BOQ 項目中缺失的數量。
      operationId: analyzeFloorPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                floor_plan_document_id:
                  type: string
                  description: 平面圖文件 ID
                boq_document_id:
                  type: string
                  description: BOQ 文件 ID
                items_to_verify:
                  type: array
                  items:
                    type: string
                  description: 需要核對的項目 ID 列表（空表示全部）
              required:
                - floor_plan_document_id
                - boq_document_id
      responses:
        '202':
          description: 核對任務已建立
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
  /api/tasks/{task_id}:
    get:
      tags:
        - Task
      summary: 取得任務狀態
      description: 查詢非同步任務的執行狀態與進度
      operationId: getTaskStatus
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '404':
          description: 找不到任務
  /api/tasks:
    get:
      tags:
        - Task
      summary: 取得任務列表
      operationId: listTasks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskStatus'
                  total:
                    type: integer
...
components:
  parameters:
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 文件 ID
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 任務 ID
  schemas:
...
    BOQItem:
      type: object
      description: |
        BOQ 項目資料模型（完全比照惠而蒙格式 15 欄）
        Excel 欄位: A:NO., B:Item no., C:Description, D:Photo, E:Dimension,
        F:Qty, G:UOM, H:Unit Rate(留空), I:Amount(留空), J:Unit CBM,
        K:Total CBM(公式), L:Note, M:Location, N:Materials/Specs, O:Brand
      required:
        - id
        - item_no
        - description
      properties:
        id:
          type: string
          format: uuid
          description: 唯一識別碼（內部使用）
        no:
          type: integer
          minimum: 1
          description: A 欄：序號
        item_no:
          type: string
          description: B 欄：項目編號
        description:
          type: string
          description: C 欄：品名描述
        photo_base64:
          type: string
          description: D 欄：圖片 Base64 編碼
          nullable: true
        dimension:
          type: string
          description: E 欄：尺寸 WxDxH (mm)
          nullable: true
        qty:
          type: number
          description: F 欄：數量
          nullable: true
        uom:
          type: string
          description: G 欄：單位（ea, m, set 等）
          nullable: true
        unit_cbm:
          type: number
          description: J 欄：單位材積
          nullable: true
        note:
          type: string
          description: L 欄：備註
          nullable: true
        location:
          type: string
          description: M 欄：位置
          nullable: true
        materials_specs:
          type: string
          description: N 欄：使用材料/規格
          nullable: true
        brand:
          type: string
          description: O 欄：品牌
          nullable: true
        source_document_id:
          type: string
          description: 來源文件 ID（內部使用）
        source_page:
          type: integer
          description: 來源頁碼（內部使用）
          nullable: true
    ExtractedImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        url:
          type: string
          description: 圖片 API URL
        format:
          type: string
        width:
          type: integer
        height:
          type: integer
        source_document_id:
          type: string
        source_page:
          type: integer
        boq_item_id:
          type: string
          nullable: true
        matched:
          type: boolean
    ParseStatistics:
      type: object
      properties:
        total_items:
          type: integer
        items_with_qty:
          type: integer
        items_with_photo:
          type: integer
        items_missing_qty:
          type: integer
        total_images:
          type: integer
        matched_images:
          type: integer
...
    TaskResponse:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        message:
          type: string
          example: 任務已建立，正在處理中
    TaskStatus:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        task_type:
          type: string
          enum: [parse_pdf, extract_images, generate_excel, analyze_floor_plan]
        status:
          type: string
          enum: [pending, processing, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
          description: 狀態訊息（繁體中文）
        result:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
...